Add Output module for server.
